Convert to java.time
Unfortunately java.time is lacking an Interval. I brought it in via a
third party library, but seeing that my goal was to create a Scala-JS
variant, I basically now have to create either my own basic idea of an
interval, or try to port threeten-extra to Scala-JS.

I also have to make some decisions about a default time zone.